gateway-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Database
asyncmy==0.2.9
aiomysql==0.2.0
sqlalchemy==2.0.23

# Redis
redis==5.0.1
hiredis==2.2.3

# Service Discovery & Configuration
python-consul==1.1.0
etcd3==0.2.0
nacos-sdk-python==1.0.0

# Monitoring & Tracing
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger==1.21.0

# Rate Limiting
slowapi==0.1.9

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-multipart==0.0.6
orjson==3.9.10

# Vault
hvac==1.2.1

